requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
pydantic-settings>=2.0
pyyaml>=6.0
langchain>=0.1.0
langchain-google-genai
langchain-groq
langchain-ollama
langchain-community
pandas>=2.0.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
requests>=2.31.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-docx>=1.1.0
pypdf>=4.0.0
unstructured>=0.12.0
torch>=2.0.0
asyncio>=3.4.3